feature / refactor : 채점 시스템 SSE → WebSocket 기반으로 전환 (#98)
* feat : WebSocket 관련 클래스, 메서드, DTO 추가

* refactor : WebSocket으로 채점 시스템 전환

+++ src/main/java/org/ezcode/codetest/application/submission/model/JudgeResult.java
@@ -7,7 +7,7 @@ public record JudgeResult(
 
     String actualOutput,
 
-    double executionTime,
+    long executionTime,
 
     long memoryUsage,
